+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * New set_new command.
 * New is_set command.
 * New unset command.
+* New array_clear command.
 * New array_contains command.
 * New map_contains_value command.
 * New map_contains_key command.